Create file to manage API in RSpec integration
