+++ rspec/progress_formatter.rb
@@ -1,6 +1,4 @@
-require 'rspec/core/formatters/progress_formatter'
-require 'uri'
-require 'net/http'
+require_relative 'api'
 
 class RSpec::Core::Formatters::ProgressFormatter
   def start(_notification)
@@ -21,36 +19,5 @@ class RSpec::Core::Formatters::ProgressFormatter
 
   def api
     @api ||= API.new
-  end
-end
-
-class API
-  attr_accessor :status
-
-  def started
-    light(:yellow)
-  end
-
-  def failed
-    light(:red)
-  end
-
-  def finished
-    light(:green) if yellow?
-  end
-
-  private
-
-  def light(color)
-    uri = uri(color)
-    Net::HTTP.new(uri.host, uri.port).get(uri.path, {})
-  end
-
-  def uri(color)
-    URI("http://localhost:3000/#{color}")
-  end
-
-  def yellow?
-    status == :yellow
   end
 end
